--- app/src/main/java/com/example/healthroutineapplication/SundayDecorator.kt
@@ -1,21 +1,0 @@
-package com.example.healthroutineapplication
-
-import android.graphics.Color
-import android.text.style.ForegroundColorSpan
-import com.prolificinteractive.materialcalendarview.CalendarDay
-import com.prolificinteractive.materialcalendarview.DayViewDecorator
-import com.prolificinteractive.materialcalendarview.DayViewFacade
-import java.util.*
-
-class SundayDecorator:DayViewDecorator {
-    private val calendar = Calendar.getInstance()
-    override fun shouldDecorate(day: CalendarDay?): Boolean {
-        day?.copyTo(calendar)
-        val weekDay = calendar.get(Calendar.DAY_OF_WEEK)
-        return weekDay==Calendar.SUNDAY
-    }
-
-    override fun decorate(view: DayViewFacade?) {
-        view?.addSpan(object : ForegroundColorSpan(Color.RED){})
-    }
-}
