+++ app/src/main/java/com/example/healthroutineapplication/activities/MemoEditActivity.kt
@@ -1,10 +1,13 @@
-package com.example.healthroutineapplication
+package com.example.healthroutineapplication.activities
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import androidx.databinding.DataBindingUtil
+import com.example.healthroutineapplication.R
+import com.example.healthroutineapplication.databases.CalendarDatabase
 import com.example.healthroutineapplication.databinding.ActivityMemoEditBinding
+import com.example.healthroutineapplication.models.CalendarDataClass
 import com.example.healthroutineapplication.models.ExerciseData
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
