+++ app/src/main/java/com/example/healthroutineapplication/activities/StepCounterActivity.kt
@@ -1,9 +1,14 @@
-package com.example.healthroutineapplication
+package com.example.healthroutineapplication.activities
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import androidx.databinding.DataBindingUtil
 import androidx.lifecycle.ViewModelProvider
+import com.example.healthroutineapplication.R
+import com.example.healthroutineapplication.StepCounterService
+import com.example.healthroutineapplication.StepCounterViewModel
+import com.example.healthroutineapplication.StepCounterViewModelFactory
+import com.example.healthroutineapplication.databases.CalendarDatabase
 import com.example.healthroutineapplication.databinding.ActivityStepCounterBinding
 
 class StepCounterActivity : AppCompatActivity() {
@@ -12,9 +17,11 @@ class StepCounterActivity : AppCompatActivity() {
     lateinit var database : CalendarDatabase
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        binding = DataBindingUtil.setContentView(this,R.layout.activity_step_counter)
+        binding = DataBindingUtil.setContentView(this, R.layout.activity_step_counter)
         database = CalendarDatabase.getInstance(application)!!
-        viewModel = ViewModelProvider(this,StepCounterViewModelFactory(database))[StepCounterViewModel::class.java]
+        viewModel = ViewModelProvider(this,
+            StepCounterViewModelFactory(database)
+        )[StepCounterViewModel::class.java]
         if(!StepCounterService.isSensor){
             binding.stepCount.text = ""
             binding.textStep.text = getString(R.string.no_sensor)
