+++ app/src/main/java/com/example/healthroutineapplication/databases/CalendarDatabase.kt
@@ -1,10 +1,12 @@
-package com.example.healthroutineapplication
+package com.example.healthroutineapplication.databases
 
 import android.content.Context
 import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
 import androidx.room.TypeConverters
+import com.example.healthroutineapplication.models.CalendarDataClass
+import com.example.healthroutineapplication.interfaces.CalendarDao
 
 @Database(entities = [CalendarDataClass::class], version = 1)
 @TypeConverters(CalendarDatabaseTypeConverters::class)
