+++ app/src/main/java/com/example/healthroutineapplication/databases/CalendarDatabaseTypeConverters.kt
@@ -1,4 +1,4 @@
-package com.example.healthroutineapplication
+package com.example.healthroutineapplication.databases
 
 import androidx.room.TypeConverter
 import com.example.healthroutineapplication.models.ExerciseData
