+++ app/src/main/java/com/example/healthroutineapplication/decorators/ExerciseDayDecorator.kt
@@ -1,10 +1,10 @@
-package com.example.healthroutineapplication
+package com.example.healthroutineapplication.decorators
 
+import com.example.healthroutineapplication.R
 import com.prolificinteractive.materialcalendarview.CalendarDay
 import com.prolificinteractive.materialcalendarview.DayViewDecorator
 import com.prolificinteractive.materialcalendarview.DayViewFacade
 import com.prolificinteractive.materialcalendarview.spans.DotSpan
-import java.text.SimpleDateFormat
 import java.util.*
