+++ app/src/main/java/com/example/healthroutineapplication/decorators/SundayDecorator.kt
@@ -1,26 +1,21 @@
-package com.example.healthroutineapplication
+package com.example.healthroutineapplication.decorators
 
 import android.graphics.Color
-import android.graphics.Typeface
 import android.text.style.ForegroundColorSpan
-import android.text.style.RelativeSizeSpan
-import android.text.style.StyleSpan
 import com.prolificinteractive.materialcalendarview.CalendarDay
 import com.prolificinteractive.materialcalendarview.DayViewDecorator
 import com.prolificinteractive.materialcalendarview.DayViewFacade
 import java.util.*
 
-class TodayDecorator:DayViewDecorator {
-    private val today = CalendarDay.today()
-
+class SundayDecorator:DayViewDecorator {
+    private val calendar = Calendar.getInstance()
     override fun shouldDecorate(day: CalendarDay?): Boolean {
-        return today == day
+        day?.copyTo(calendar)
+        val weekDay = calendar.get(Calendar.DAY_OF_WEEK)
+        return weekDay==Calendar.SUNDAY
     }
 
     override fun decorate(view: DayViewFacade?) {
-        view?.apply {
-            addSpan(object : StyleSpan(Typeface.BOLD){})
-            addSpan(object : RelativeSizeSpan(1.2f){})
-        }
+        view?.addSpan(object : ForegroundColorSpan(Color.RED){})
     }
 }
