+++ app/src/main/java/com/example/healthroutineapplication/models/CalendarDataClass.kt
@@ -1,4 +1,4 @@
-package com.example.healthroutineapplication
+package com.example.healthroutineapplication.models
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
