+++ app/src/main/res/layout/activity_exercise_routine.xml
@@ -5,7 +5,7 @@
 <androidx.constraintlayout.widget.ConstraintLayout
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ExerciseRoutineActivity">
+    tools:context=".activities.ExerciseRoutineActivity">
 
     <TextView
         android:id="@+id/exerciseRoutineTitle"
