+++ app/src/main/res/layout/activity_exercise_starting.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ExerciseStartingActivity">
+    tools:context=".activities.ExerciseStartingActivity">
     <TextView
         android:id="@+id/exerciseState"
         android:layout_width="match_parent"
