+++ app/src/main/res/layout/activity_step_counter.xml
@@ -5,7 +5,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".StepCounterActivity">
+        tools:context=".activities.StepCounterActivity">
 
         <TextView
             android:id="@+id/textView"
